// frontend/src/types/userTypes.ts
import { LoginInputState, SignupInputState } from "@/schema/userSchema";

export type User = {
	fullname: string;
	email: string;
	contact: number;
	address: string;
	city: string;
	country: string;
	profilePicture: string;
	admin: boolean;
	isVerified: boolean;
}

export type UserState = {
	user: User | null;
	isAuthenticated: boolean;
	isCheckingAuth: boolean;
	loading: boolean; 
	signup: (input: SignupInputState) => Promise<void>;
	login: (input: LoginInputState) => Promise<void>;
	verifyEmail: (verificationCode: string) => Promise<void>;
	checkAuthentication: () => Promise<void>;
	logout: () => Promise<void>;
	forgotPassword: (email: string) => Promise<void>;
	resetPassword: (token: string, newPassword: string) => Promise<void>;
	updateProfile: (input: any) => Promise<void>; 
}